IDEA-128864 diff: fix invalid range applying
MergeOperation is created with fixed TextRange, so it will not be updated on change, like RangeHighlighters (and usage of RangeHighlighter is dangerous, as it could be updated incorrectly)
So we have to disable all actions until diff rebuild to prevent document corruption.

The 'right' fix of this issue could be to store line numbers inside action, update them on document change (like in LST), and use Document.getLineStartOffset() and etc inside write action.